@@ -11,48 +11,40 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^6.1.10",
-    "@angular/common": "^6.0.3",
-    "@angular/compiler": "^6.0.3",
-    "@angular/core": "^6.0.3",
-    "@angular/forms": "^6.0.3",
-    "@angular/http": "^6.0.3",
-    "@angular/platform-browser": "^6.0.3",
-    "@angular/platform-browser-dynamic": "^6.0.3",
-    "@angular/router": "^6.0.3",
-    "bootstrap": "^4.6.0",
-    "core-js": "^2.5.4",
-    "jquery": "^3.6.0",
+    "@angular/animations": "^11.2.8",
+    "@angular/common": "~11.2.8",
+    "@angular/compiler": "~11.2.8",
+    "@angular/core": "~11.2.8",
+    "@angular/forms": "~11.2.8",
+    "@angular/platform-browser": "~11.2.8",
+    "@angular/platform-browser-dynamic": "~11.2.8",
+    "@angular/router": "~11.2.8",
     "jw-angular-pagination": "^1.2.0",
     "ng-block-ui": "^3.0.2",
-    "ngx-bootstrap": "^2.0.5",
-    "ngx-toastr": "^10.1.0",
-    "popper.js": "^1.16.1",
-    "rxjs": "6.0.0",
-    "rxjs-compat": "6.2.2",
+    "ngx-bootstrap": "^6.2.0",
+    "ngx-toastr": "^13.2.1",
+    "rxjs": "~6.6.0",
     "sweetalert2": "^10.15.7",
-    "zone.js": "^0.8.26"
+    "tslib": "^2.0.0",
+    "zone.js": "~0.11.3"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.6.8",
-    "@angular/cli": "~6.0.8",
-    "@angular/compiler-cli": "^6.0.3",
-    "@angular/language-service": "^6.0.3",
-    "@types/jasmine": "~2.8.6",
-    "@types/jasminewd2": "~2.0.3",
-    "@types/jquery": "^3.5.5",
-    "@types/node": "~8.9.4",
-    "codelyzer": "~4.2.1",
-    "jasmine-core": "~2.99.1",
-    "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~1.7.1",
-    "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "~2.0.0",
-    "karma-jasmine": "~1.1.1",
-    "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "~5.3.0",
-    "ts-node": "~5.0.1",
-    "tslint": "~5.9.1",
-    "typescript": "~2.7.2"
+    "@angular-devkit/build-angular": "~0.1102.7",
+    "@angular/cli": "~11.2.7",
+    "@angular/compiler-cli": "~11.2.8",
+    "@types/jasmine": "~3.6.0",
+    "@types/node": "^12.11.1",
+    "codelyzer": "^6.0.0",
+    "jasmine-core": "~3.6.0",
+    "jasmine-spec-reporter": "~5.0.0",
+    "karma": "~6.1.0",
+    "karma-chrome-launcher": "~3.1.0",
+    "karma-coverage": "~2.0.3",
+    "karma-jasmine": "~4.0.0",
+    "karma-jasmine-html-reporter": "^1.5.0",
+    "protractor": "~7.0.0",
+    "ts-node": "~8.3.0",
+    "tslint": "~6.1.0",
+    "typescript": "~4.1.5"
   }
 }
