@@ -1,4 +1,4 @@
-import { Component } from "@angular/core";
+import { Component, OnInit } from '@angular/core';
 import { NgForm } from "@angular/forms";
 import { Router } from "@angular/router";
 import { AuthService } from "../Service/auth.service";
@@ -6,38 +6,49 @@ import { ToastrService } from 'ngx-toastr';
 import { BlockUI, NgBlockUI } from 'ng-block-ui';
 
 @Component({
-  selector: 'login',
-  templateUrl: './Login.component.html'
+  selector: 'app-login',
+  templateUrl: './login.component.html'
 })
-export class loginComponent {
+export class LoginComponent implements OnInit {
 
+  @BlockUI() blockUI: NgBlockUI;
   title = 'Login';
   public username: string;
   public password: string;
   public errorMessage: string;
   submit:boolean = false;
-  @BlockUI() blockUI : NgBlockUI;
 
   constructor(private router: Router, private auth: AuthService,private toastr: ToastrService) { }
   
   authenticate(form: NgForm) {
+    console.log("test")
     if (form.valid) {
-      this.blockUI.start("loading..");
+      this.blockUI.start('Loading...'); // Start blocking
+ 
+      setTimeout(() => {
+        this.blockUI.stop(); // Stop blocking
+      },500);
       this.auth.authenticate(this.username, this.password)
         .subscribe(response => {                    
           if (response) {
-            this.router.navigateByUrl("/Home");
+            this.router.navigateByUrl("/Task/Dashboard");
             //location.reload();
-            this.toastr.success("Successfully Log In");
+            //alert("Successfully Log In");            
+            this.toastr.success('Login Successfull!', 'Toastr fun!');
+            this.blockUI.stop();
           }
           this.errorMessage = "Authentication Failed";
-          this.blockUI.stop();
+
         },error =>{
-          this.toastr.error(error.message);
+          this.toastr.warning('Incorrect username or password!', 'Toastr fun!');
           this.blockUI.stop();
         })
     } else {
       this.errorMessage = "Form Data Invalid";
     }
   }
+
+  ngOnInit(): void {
+  }
+
 }
