@@ -1,18 +1,20 @@
-import { Injectable } from "@angular/core";
-import { ActivatedRouteSnapshot, RouterStateSnapshot,  Router} from "@angular/router";
-import { AuthService } from "./auth.service";
+import { Injectable } from '@angular/core';
+import { ActivatedRouteSnapshot, Router, RouterStateSnapshot } from '@angular/router';
+import { AuthService } from './auth.service';
 
-@Injectable()
+@Injectable({
+  providedIn: 'root'
+})
 export class AuthGuard {
 
-    constructor(private router: Router,  private auth: AuthService) { }
+  constructor(private router: Router,  private auth: AuthService) { }
 
-    canActivate(route: ActivatedRouteSnapshot,
-        state: RouterStateSnapshot): boolean {
-        if (!this.auth.authenticated) {
-            this.router.navigateByUrl("/login");
-            return false;
-        }
-        return true;
-    }
+  canActivate(route: ActivatedRouteSnapshot,
+      state: RouterStateSnapshot): boolean {
+      if (!this.auth.authenticated) {
+          this.router.navigateByUrl("/login");
+          return false;
+      }
+      return true;
+  }
 }
